% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFcsv2Rdata.R
\name{DFcsv2Rdata}
\alias{DFcsv2Rdata}
\title{Convert CSV file to Rdata}
\usage{
DFcsv2Rdata(File_path)
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt).}
}
\value{
Rdata
}
\description{
This function convert the CSV file gained through the
PurseNT functions to Rdata. Then it check the data table.
}
\examples{
\dontrun{

#rdfs:label
File_path <- "./Label_en_rdfs.nt"

#Run
PurseNT_Label(File_path)
DFcsv2Rdata(File_path)

}

}
\author{
Satoshi Kume
}
