% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFcsv2Rds.R
\name{DFcsv2Rds}
\alias{DFcsv2Rds}
\title{Convert CSV file to Rds file}
\usage{
DFcsv2Rds(File_path, Type)
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt).}
}
\value{
Rds
}
\description{
This function convert the CSV file gained through the
PurseNT functions to Rds. Then it check the data table.
}
\examples{
\dontrun{

#rdfs:label
File_path <- "./Label_en_rdfs.nt"

#Run
PurseNT_Label(File_path)
DFcsv2Rds(File_path)

}

}
\author{
Satoshi Kume
}
