% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFcsv2SQLite.R
\name{DFcsv2SQLite}
\alias{DFcsv2SQLite}
\title{Convert CSV file to SQLite}
\usage{
DFcsv2SQLite(File_path, SQLite_name)
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt).}

\item{SQLite_name}{}
}
\value{
Rdata
}
\description{
This function convert the CSV file gained through the
PurseNT functions to Rdata. Then it check the data table.
}
\examples{
\dontrun{

#rdfs:label
File_path <- "./Label_en_rdfs.nt"

#Run
PurseNT_Label(File_path)
DFcsv2SQLite(File_path)

}

}
\author{
Satoshi Kume
}
