% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectPurseNT.R
\name{DirectPurseNT}
\alias{DirectPurseNT}
\title{Directly convert NT file to each file}
\usage{
DirectPurseNT(File_path, Type = "Wikidata")
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt).}

\item{Type}{a type of dataset; select only Wikidata, CHEBI, or Mesh.}
}
\value{
Rds
}
\description{
This function convert the TSV file gained through the
PurseNT functions to Rds. Then it check the data table.
}
\examples{
\dontrun{
#Not Run
#rdfs:label
File_path <- "./Label_en_rdfs.nt"

#Run
PurseNT_Label(File_path, tsv=TRUE)
DFtsv2Rds(File_path)

}

}
\author{
Satoshi Kume
}
