% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mesh_PurseNT.R
\name{Mesh_PurseNT}
\alias{Mesh_PurseNT}
\title{Convert the triples witout the label information to the CSV file for Mesh dump}
\usage{
Mesh_PurseNT(
  File_path,
  Subject = c("http://id.nlm.nih.gov/mesh/2019/", "mesh2019:",
    "http://id.nlm.nih.gov/mesh/2020/", "mesh2020:", "http://id.nlm.nih.gov/mesh/2021/",
    "mesh2021:", "http://id.nlm.nih.gov/mesh/", "mesh:"),
  Property = c("http://id.nlm.nih.gov/mesh/vocab#", "meshv:",
    "http://www.w3.org/1999/02/22-rdf-syntax-ns#", "rdf:",
    "http://www.w3.org/2000/01/rdf-schema#", "rdfs:"),
  Object = c("http://id.nlm.nih.gov/mesh/2019/", "mesh2019:",
    "http://id.nlm.nih.gov/mesh/2020/", "mesh2020:", "http://id.nlm.nih.gov/mesh/2021/",
    "mesh2021:", "http://id.nlm.nih.gov/mesh/", "mesh:",
    "http://id.nlm.nih.gov/mesh/vocab#", "meshv:")
)
}
\arguments{
\item{File_path}{a character vector,
indicating a N-triple (NT) file path (.nt).}
}
\value{
CSV file
}
\description{
This function use the NT file and
convert from the Mesh dump to CSV file.
}
\examples{
\dontrun{

#Class hierarchy
File_path <- "./mesh2019_others2.nt"

# Run
Mesh_PurseNT(File_path)
DFcsv2Rdata(File_path)

}

}
\author{
Satoshi Kume
}
