% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mesh_PurseNT_Label.R
\name{Mesh_PurseNT_Label}
\alias{Mesh_PurseNT_Label}
\title{Convert labels to CSV file for Mesh dump}
\usage{
Mesh_PurseNT_Label(File_path)
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt).}
}
\value{
CSV file
}
\description{
This function convert the NT file from the Mesh dump
to CSV file.
}
\examples{
\dontrun{

#rdfs:label
File_path <- "./Label_en_rdfs.nt"

#Run
Mesh_PurseNT_Label(File_path)
DFcsv2Rdata(File_path)

}

}
\author{
Satoshi Kume
}
