% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse4Mesh.R
\name{Parse4Mesh}
\alias{Parse4Mesh}
\title{Parse Mesh data to a new class hierarchy}
\usage{
Parse4Mesh(File_path, Labels = FALSE)
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt).}

\item{Labels}{a logical}
}
\description{
This function convert to a new class hierarchy.
}
\examples{
\dontrun{

#Run
File_path <- "./Label_en_rdfs.nt"
Parse4Mesh(File_path)

}

}
\author{
Satoshi Kume
}
