% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseNT.R
\name{ParseNT}
\alias{ParseNT}
\title{Convert the triples  to the CSV or TSV file}
\usage{
ParseNT(File_path, tsv = TRUE)
}
\arguments{
\item{File_path}{a character vector,
indicating a N-triple (NT) file path (.nt).}

\item{tsv}{save as TSV format}
}
\value{
CSV or TSV file
}
\description{
This function use the NT file and
convert from the small dump dataset (less than nearly 2GB) to CSV or TSV file.
}
\examples{
\dontrun{

#Class hierarchy
File_path <- "./mesh2019_others2.nt"

# Run
ParseNT(File_path)
DFcsv2Rdata(File_path)

}

}
\author{
Satoshi Kume
}
