% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcCSV_af_nkf.R
\name{ProcCSV_af_nkf}
\alias{ProcCSV_af_nkf}
\title{Processing after the conversion to the Multi-Byte Character.}
\usage{
ProcCSV_af_nkf(File_path)
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt) or CSV file.}
}
\value{
Rds
}
\description{
This function provide a Processing after the conversion from
their unicode (Escape character) to Multi-Byte Character (Japanese).
}
\examples{
\dontrun{

File_path <- "./Label_en_rdfs.nt"
ProcCSV_af_nkf(File_path)

}

}
\author{
Satoshi Kume
}
