% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{ProxySet}
\alias{ProxySet}
\title{Proxy settings}
\usage{
ProxySet(Locations)
}
\arguments{
\item{Locations}{a character vector}
}
\description{
This function provides the proxy settings.
}
\examples{
\dontrun{

#Class hierarchy
File_path <- "./mesh2019_others2.nt"

# Run
PurseNT(File_path)
DFcsv2Rdata(File_path)

}

}
\author{
Satoshi Kume
}
