% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wikidata_ParseNT_others.R
\name{Wikidata_ParseNT_Others}
\alias{Wikidata_ParseNT_Others}
\title{Convert the triples without class hierarchy and labels for Wikidata dump}
\usage{
Wikidata_ParseNT_Others(File_path)
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt).}
}
\value{
CSV file
}
\description{
This function convert the NT file from the
Wikidata dump to CSV file.
}
\examples{
\dontrun{

#Class hierarchy
File_path <- "./Relation_KEGG_ID.nt"

#Run
Wikidata_ParseNT_Others(File_path)
DFcsv2Rds(File_path)

}

}
\author{
Satoshi Kume
}
