% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mesh_PurseNT_ClassHierarchy.R
\name{Mesh_PurseNT}
\alias{Mesh_PurseNT}
\title{Convert the class hierarchy to CSV file for Mesh dump}
\usage{
Mesh_PurseNT(
  File_path,
  Subject = c("http://id.nlm.nih.gov/mesh/2019/", "mesh2019:"),
  Property = c("http://id.nlm.nih.gov/mesh/vocab#", "meshv:"),
  Object = c("http://id.nlm.nih.gov/mesh/2019/", "mesh2019:")
)
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt).}
}
\value{
CSV file
}
\description{
This function convert the NT file
from the Mesh dump to CSV file.
}
\examples{
\dontrun{

#Class hierarchy
File_path <- "./mesh2019_others2.nt"

# Run
Mesh_PurseNT(File_path)
DFcsv2Rdata(File_path)

}

}
\author{
Satoshi Kume
}
