% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurseNT.R
\name{PurseNT}
\alias{PurseNT}
\title{Convert the triples  to the CSV file}
\usage{
PurseNT(File_path)
}
\arguments{
\item{File_path}{a character vector,
indicating a N-triple (NT) file path (.nt).}
}
\value{
CSV file
}
\description{
This function use the NT file and
convert from the small dump dataset (less than nearly 2GB) to CSV file.
}
\examples{
\dontrun{

#Class hierarchy
File_path <- "./mesh2019_others2.nt"

# Run
PurseNT(File_path)
DFcsv2Rdata(File_path)

}

}
\author{
Satoshi Kume
}
