% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurseNT_Label.R
\name{PurseNT_Label}
\alias{PurseNT_Label}
\title{Convert labels to CSV file}
\usage{
PurseNT_Label(
  File_path,
  Subject = c("http://id.nlm.nih.gov/mesh/2019/", "mesh2019:",
    "http://id.nlm.nih.gov/mesh/2020/", "mesh2020:", "http://id.nlm.nih.gov/mesh/2021/",
    "mesh2021:", "http://id.nlm.nih.gov/mesh/", "mesh:"),
  Property = c("http://id.nlm.nih.gov/mesh/vocab#", "meshv:",
    "http://www.w3.org/1999/02/22-rdf-syntax-ns#", "rdf:",
    "http://www.w3.org/2000/01/rdf-schema#", "rdfs:")
)
}
\arguments{
\item{File_path}{a character vector for a N-triple (NT) file (.nt).}
}
\value{
CSV file
}
\description{
This function convert the NT file from the Mesh dump
to CSV file.
}
\examples{
\dontrun{

#rdfs:label
File_path <- "./Label_en_rdfs.nt"

#Run
PurseNT_Label(File_path)
DFcsv2Rdata(File_path)

}

}
\author{
Satoshi Kume
}
